Move the data-generation script inside the unit test file & remove db.py update unit test
